+++ sdk/schemaregistry/azure-schemaregistry-avroencoder/azure/schemaregistry/encoder/avroencoder/_abstract_avro_encoder.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-from typing import BinaryIO, TypeVar, Union, Optional
+from typing import BinaryIO, TypeVar, Union, Optional, Any
 from abc import abstractmethod
 
 ObjectType = TypeVar("ObjectType")
@@ -51,9 +51,7 @@ class AbstractAvroObjectEncoder(object):
     def decode(
         self,
         content: Union[bytes, BinaryIO],
-        schema: str,
-        *,
-        readers_schema: Optional[str]
+        reader: Any
     ):
         """Read the binary representation into a specific type.
         Return type will be ignored, since the schema is deduced from the provided bytes.