+++ sdk/schemaregistry/azure-schemaregistry-avroencoder/tests/test_avro_encoder.py
@@ -29,7 +29,7 @@ import json
 
 from azure.schemaregistry import SchemaRegistryClient
 from azure.schemaregistry.encoder.avroencoder import AvroEncoder
-from azure.schemaregistry.encoder.avroencoder import AvroEncodeError
+from azure.schemaregistry.encoder.avroencoder import AvroEncodeError, InvalidSchemaError
 from devtools_testutils import AzureRecordedTestCase, EnvironmentVariableLoader, recorded_by_proxy, AzureTestCase
 
 import avro
@@ -58,7 +58,8 @@ class TestAvroEncoder(AzureRecordedTestCase):
         encoded_payload = raw_avro_object_encoder.encode(dict_content, schema_str)
 
         # decoding part
-        decoded_content = raw_avro_object_encoder.decode(encoded_payload, schema_str)
+        reader = raw_avro_object_encoder.get_schema_reader(schema_str)
+        decoded_content = raw_avro_object_encoder.decode(encoded_payload, reader)
 
         assert decoded_content["name"] == u"Ben"
         assert decoded_content["favorite_number"] == 7
@@ -66,7 +67,8 @@ class TestAvroEncoder(AzureRecordedTestCase):
 
         dict_content_missing_optional_fields = {"name": u"Alice"}
         encoded_payload = raw_avro_object_encoder.encode(dict_content_missing_optional_fields, schema_str)
-        decoded_content = raw_avro_object_encoder.decode(encoded_payload, schema_str)
+        reader = raw_avro_object_encoder.get_schema_reader(schema_str)
+        decoded_content = raw_avro_object_encoder.decode(encoded_payload, reader)
 
         assert decoded_content["name"] == u"Alice"
         assert not decoded_content["favorite_number"]
@@ -122,12 +124,13 @@ class TestAvroEncoder(AzureRecordedTestCase):
             def __init__(self, not_content):
                 self.not_content = not_content
 
-        with pytest.raises(AvroEncodeError) as e:    # caught avro AvroEncodeError
+        with pytest.raises(AvroEncodeError) as e:    # caught AvroEncodeError
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_str, message_type=BadExample) 
         assert "subtype of the MessageType" in (str(e.value))
+        assert "message_content" in e.value.details
 
         bad_ex = BadExample('fake')
-        with pytest.raises(AvroEncodeError) as e:    # caught avro AvroEncodeError
+        with pytest.raises(TypeError) as e:    # caught TypeError
             sr_avro_encoder.decode(message=bad_ex) 
         assert "subtype of the MessageType" in (str(e.value))
 
@@ -263,7 +266,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "fields":[{"name":"name","type":"string"}]
         }
         invalid_schema_string = "{}".format(invalid_schema)
-        with pytest.raises(AvroEncodeError):    # caught avro AvroEncodeError
+        with pytest.raises(InvalidSchemaError):    # caught avro InvalidSchemaError
             sr_avro_encoder.encode({"name": u"Ben"}, schema=invalid_schema_string) 
 
     ######################### PRIMITIVES #########################
@@ -277,7 +280,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
         sr_avro_encoder = AvroEncoder(client=sr_client, group_name=schemaregistry_group, auto_register=True)
 
         primitive_string = "string"
-        with pytest.raises(AvroEncodeError) as e:
+        with pytest.raises(InvalidSchemaError) as e:
             sr_avro_encoder.encode("hello", schema=primitive_string) 
 
     ######################### type fixed #########################
@@ -295,19 +298,19 @@ class TestAvroEncoder(AzureRecordedTestCase):
         #sr_avro_encoder.encode({}, schema=fixed_type_ignore_logical_type_error) 
 
         schema_no_size = """{"type": "fixed", "name":"User"}"""
-        with pytest.raises(AvroEncodeError):    # caught AvroException
+        with pytest.raises(InvalidSchemaError):    # caught AvroException
             sr_avro_encoder.encode({}, schema=schema_no_size) 
 
         schema_no_name = """{"type": "fixed", "size": 3}"""
-        with pytest.raises(AvroEncodeError):    # caught AvroEncodeError
+        with pytest.raises(InvalidSchemaError):    # caught InvalidSchemaError
             sr_avro_encoder.encode({}, schema=schema_no_name) 
 
         schema_wrong_name = """{"type": "fixed", "name": 1, "size": 3}"""
-        with pytest.raises(AvroEncodeError):    # caught AvroEncodeError
+        with pytest.raises(InvalidSchemaError):    # caught InvalidSchemaError
             sr_avro_encoder.encode({}, schema=schema_wrong_name) 
 
         schema_wrong_namespace = """{"type": "fixed", "name": "User", "size": 3, "namespace": 1}"""
-        with pytest.raises(AvroEncodeError):    # caught AvroEncodeError
+        with pytest.raises(InvalidSchemaError):    # caught InvalidSchemaError
             sr_avro_encoder.encode({}, schema=schema_wrong_namespace) 
 
     ######################### type unspecified #########################
@@ -325,7 +328,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "namespace":"example.avro",
             "fields":[{"name":"name","type":"string"}]
         }"""
-        with pytest.raises(AvroEncodeError):    # caught avro AvroEncodeError
+        with pytest.raises(InvalidSchemaError):    # caught avro InvalidSchemaError
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_no_type) 
 
         schema_wrong_type_type = """{
@@ -334,7 +337,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "namespace":"example.avro",
             "fields":[{"name":"name","type":"string"}]
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_wrong_type_type) 
 
     ######################### RECORD SCHEMA #########################
@@ -381,7 +384,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "namespace":"9example.avro",
             "fields":[{"name":"name","type":"string"}]
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_invalid_fullname) 
 
         schema_invalid_name_in_fullname = """{
@@ -389,7 +392,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "type":"record",
             "fields":[{"name":"name","type":"string"}]
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_invalid_name_in_fullname) 
 
         schema_invalid_name_reserved_type = """{
@@ -397,7 +400,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "type":"record",
             "fields":[{"name":"name","type":"string"}]
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_invalid_name_reserved_type)
 
         schema_wrong_type_name = """{
@@ -406,7 +409,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "namespace":"example.avro",
             "fields":[{"name":"name","type":"string"}]
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_wrong_type_name) 
 
         schema_no_name = """{
@@ -414,7 +417,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "type":"record",
             "fields":[{"name":"name","type":"string"}]
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_no_name) 
 
     @SchemaRegistryEnvironmentVariableLoader()
@@ -450,7 +453,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "namespace":"example.avro",
             "type":"record"
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_no_fields) 
 
         schema_wrong_type_fields = """{
@@ -459,7 +462,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "type":"record"
             "fields": "hello"
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_wrong_type_fields) 
 
         schema_wrong_field_item_type = """{
@@ -468,7 +471,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "type":"record"
             "fields": ["hello"]
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_wrong_field_item_type) 
 
         schema_record_field_no_name= """{
@@ -477,7 +480,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "type":"record",
             "fields":[{"type":"string"}]
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_record_field_no_name) 
 
         schema_record_field_wrong_type_name= """{
@@ -486,7 +489,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "type":"record",
             "fields":[{"name": 1, "type":"string"}]
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_record_field_wrong_type_name) 
 
         schema_record_field_with_invalid_order = """{
@@ -495,7 +498,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "type":"record",
             "fields":[{"name":"name","type":"string","order":"fake_order"}]
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_record_field_with_invalid_order) 
 
         schema_record_duplicate_fields = """{
@@ -504,7 +507,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "type":"record",
             "fields":[{"name":"name","type":"string"}, {"name":"name","type":"string"}]
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_record_duplicate_fields) 
 
         schema_field_type_invalid = """{
@@ -513,7 +516,7 @@ class TestAvroEncoder(AzureRecordedTestCase):
             "type":"record",
             "fields":[{"name":"name","type":1}]
         }"""
-        with pytest.raises(AvroEncodeError):
+        with pytest.raises(InvalidSchemaError):
             sr_avro_encoder.encode({"name": u"Ben"}, schema=schema_field_type_invalid) 
 
     #################################################################